chore: harden linting for attack log
