+++ src/features/attack-log/index.ts
@@ -1,3 +1,4 @@
 export { AttackLogPanel } from './AttackLogPanel';
-export { AttackLogProvider, useAttackLogContext } from './context';
+export { AttackLogProvider } from './context';
+export { useAttackLogContext } from './AttackLogContext';
 export { AttackLogActions } from './components/AttackLogActions';
